flask_auth/requirements.txt
flask==2.2.2
SQLAlchemy==1.4.45
redis==4.4.0
hiredis==2.1.0 #  Типо ускорят парсинг ответа от Редиса.
# Flask-SQLAlchemy==3.0.2
psycopg2-binary==2.9.5
pydantic==1.10.2
python-dotenv==0.21.0
flask-jwt-extended==4.4.4
gunicorn==20.1.0
flask-swagger-ui==4.11.1
apispec[marshmallow]==6.0.2
apispec-webframeworks==0.5.2
marshmallow==3.19.0
gevent
alembic==1.9.1
Flask-Migrate==4.0.2 # Да, так миграции попроще).
# Flask_Security==3.0.0
Authlib==1.2.0
requests==2.28.2
# opentelemetry-api==1.10.0
# opentelemetry-sdk==1.10.0
# opentelemetry-instrumentation-flask==0.29b1
# opentelemetry-exporter-jaeger==1.10.0
python-logstash==0.4.8
aiohttp==3.8.3
requests==2.28.2
pika==1.3.1